fix(spa): align isOpeningBalancesAccount with Go + consolidate account API in lib/accounts.ts
